+++ src/main/kotlin/net/fallingangel/jimmercraft/facts/Facet.kt
@@ -1,0 +1,5 @@
+package net.fallingangel.jimmercraft.facts
+
+interface Facet<K, V> {
+    operator fun get(key: K): V
+}