+++ src/main/kotlin/net/fallingangel/jimmercraft/facts/JimmerFacts.kt
@@ -1,0 +1,7 @@
+package net.fallingangel.jimmercraft.facts
+
+object JimmerFacts {
+    operator fun <K, V> get(facet: Facet<K, V>, key: K): V {
+        return facet[key]
+    }
+}