+++ src/main/kotlin/net/fallingangel/jimmercraft/rule/AnnotationValue.kt
@@ -1,0 +1,5 @@
+package net.fallingangel.jimmercraft.rule
+
+import com.intellij.psi.PsiElement
+
+data class AnnotationValue<T>(val value: T, val psi: PsiElement)