+++ src/main/kotlin/net/fallingangel/jimmercraft/rule/Rule.kt
@@ -1,0 +1,5 @@
+package net.fallingangel.jimmercraft.rule
+
+interface Rule {
+    operator fun invoke(site: PropAnnotationSite): List<Diagnostic>
+}